scene/camera: file movements to remove scene.h
With this change, each book how has a camera.h file which holds both the
camera class definition and the legacy scene class definition.

This is in preparation to retire the scene class and move all scene
functionality into the camera class.
