+++ src/InOneWeekend/main.cc
@@ -15,7 +15,6 @@
 #include "color.h"
 #include "hittable_list.h"
 #include "material.h"
-#include "scene.h"
 #include "sphere.h"
