+++ src/TheRestOfYourLife/main.cc
@@ -17,7 +17,6 @@
 #include "hittable_list.h"
 #include "material.h"
 #include "quad.h"
-#include "scene.h"
 #include "sphere.h"
